move evaluateFhirDataService into view-data
