+++ containers/ecr-viewer/src/app/view-data/components/EcrDocument/accordion-items.tsx
@@ -2,15 +2,6 @@ import React from "react";
 
 import { Bundle } from "fhir/r4";
 
-import {
-  evaluateDemographicsData,
-  evaluateSocialData,
-  evaluateEncounterData,
-  evaluateProviderData,
-  evaluateFacilityData,
-  evaluateHospitalEncounterData,
-  evaluatePregnancyData,
-} from "@/app/services/evaluateFhirDataService";
 import { evaluateLabInfoData } from "@/app/services/labsService";
 import { AccordionItem } from "@/app/types";
 import { evaluateAll } from "@/app/utils/evaluate";
@@ -25,6 +16,15 @@ import PregnancyInfo from "@/app/view-data/components/PregnancyInfo";
 import SocialHistory from "@/app/view-data/components/SocialHistory";
 import UnavailableInfo from "@/app/view-data/components/UnavailableInfo";
 import { evaluateEcrMetadata } from "@/app/view-data/services/ecrMetadataService";
+import {
+  evaluateDemographicsData,
+  evaluateSocialData,
+  evaluateEncounterData,
+  evaluateProviderData,
+  evaluateFacilityData,
+  evaluateHospitalEncounterData,
+  evaluatePregnancyData,
+} from "@/app/view-data/services/evaluateFhirDataService";
 
 import { evaluateClinicalData } from "./clinical-data";
