+++ containers/ecr-viewer/src/app/view-data/page.tsx
@@ -3,10 +3,6 @@ import React from "react";
 import { notFound } from "next/navigation";
 
 import { GenericError, RetrievalFailed } from "@/app/components/ErrorPage";
-import {
-  evaluatePatientDOB,
-  evaluatePatientName,
-} from "@/app/services/evaluateFhirDataService";
 import { getFhirData, isSuccessResponse } from "@/app/services/fhirDataService";
 import { isLoggedInUserEcrAuthed } from "@/app/services/userService";
 import { getLoggedInUserSession } from "@/app/utils/auth-utils";
@@ -21,6 +17,10 @@ import {
   evaluateEcrSummaryEncounterDetails,
   evaluateEcrSummaryPatientDetails,
 } from "./services/ecrSummaryService";
+import {
+  evaluatePatientDOB,
+  evaluatePatientName,
+} from "./services/evaluateFhirDataService";
 
 /**
  * Functional component for rendering the eCR Viewer page.
